batch: Extract source cache persistence
The batch package records source commits in the session mapping while
source snapshot construction still builds the commits that those entries
point to.

The project does not provide a named home for the mapping itself. Code that
only needs to read or write the session cache must import the snapshot
module, which makes the cache look like part of commit construction.

This commit addresses that by moving the session source-cache helpers into
batch.source_cache and by updating batch persistence callers to import that
module directly.

Subsequent test commits will pin direct cache imports and guard the final
boundary.
